refactored expander and ref resolver for readability
[NO CHANGE IN FUNCTIONALITY]

* refactored associated units tests for readability
* reshuffled resolver funcs and associated tests to a dedicated file
* removed panic() in ResolveRef (now return error) [used by go-swagger]
