+++ go.mod
@@ -5,7 +5,7 @@ require (
 	github.com/go-openapi/jsonreference v0.19.5
 	github.com/go-openapi/swag v0.19.12
 	github.com/stretchr/testify v1.6.1
-	golang.org/x/net v0.0.0-20201110031124-69a78807bb2b // indirect
+	golang.org/x/net v0.0.0-20201224014010-6772e930b67b // indirect
 	golang.org/x/text v0.3.4 // indirect
 	gopkg.in/yaml.v2 v2.4.0
 )
